Changes to the branch moving things in directories and starting Navbar
