+++ resume-app/src/index.css
@@ -1,20 +1,48 @@
   .bnwUs{
     background-image: url(./images/bnwUs.jpg);
-    /* width: 4vw; */
-    /* height: 5vh; */
-    width: 200px;
-    height: 400px
+    width: 25vw;
+    height: 65vh;
+    
   }  
 
-
-
-.homeClass{
+.nav-bar-outer{
+  background-color: grey;
+  /* min-height: 100vh; */
+  display: flex;
+  flex-direction: row;
+  align-items: inherit;
+  justify-content: center;
+  color: black;
+}
+.nav-bar{
+  list-style-type: none;
+  display: inline-flex;
+}
+.nav-items{
+  padding-top: 1vh;
+  margin-left: 3vw;
+}
+.nav-buttons{
   background-color: #282c34;
+  border-radius: 50%;
+}
+.nav-expand{
+  justify-content: left;
+}
+.nav-buttons:hover{
+  color: whitesmoke;
+}
+
+
+
+  
+.homeClass{
+  /* background-color: #282c34;
   min-height: 100vh;
   display: flex;
   flex-direction: row;
   align-items: center;
   justify-content: center;
-  color: whitesmoke;
+  color: whitesmoke; */
 }
