+++ resume-app/src/index.js
@@ -4,8 +4,9 @@ import './index.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 // import App from './App';
-import reportWebVitals from './reportWebVitals';
+
 import Routing from './Routing.tsx';
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -14,7 +15,4 @@ root.render(
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+
