+++ resume-app/src/types/index.d.ts
@@ -1,0 +1,2 @@
+declare module '*.jpg';
+declare module '*.png';
